pyproject.toml
[tool.poetry]
name = "discord_bot"
version = "0.1.0"
description = ""
authors = ["Rafhael Payão <rafaeltmbr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
start = "discord_bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
